Avancez del Controlador UsuarioController
Action:
- Index
- Details
- Create
- Edit
- Delete

+++ MobileNic/MobileNic.MVC/Models/Usuario.cs
@@ -65,7 +65,7 @@ namespace MobileNic.MVC.Models
 
         [Required]
         [Display(Name = "Municipio")]
-        public int IdMunicipio { get; set; }
+        public short IdMunicipio { get; set; }
     }
 
     public partial class Usuario
